+++ docker/db/mysql/init/init.sql
@@ -1,12 +1,12 @@
 
 
-create table piece (
+create table board (
     position varchar(10) not null,
     name varchar(20) not null,
     primary key(position)
 );
 
-create table board (
+create table chessgame (
     id int(5) primary key auto_increment,
     state varchar(20) not null
 );
