+++ src/main/java/chess/controller/ConsoleController.java
@@ -1,0 +1,2 @@
+package chess.controller;public class ConsoleController {
+}
