+++ src/main/java/chess/service/ChessService.java
@@ -1,0 +1,2 @@
+package chess.service;public class ChessService {
+}
