+++ src/test/java/chess/dao/BoardDaoTest.java
@@ -1,33 +1,37 @@
 package chess.dao;
 
-import chess.domain.Team;
-import chess.domain.state.turn.State;
-import chess.domain.state.turn.WhiteTurn;
+import chess.domain.board.Board;
+import chess.domain.board.BoardBuilder;
+import chess.domain.board.Position;
+import chess.domain.piece.Piece;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-public class BoardDaoTest {
+public class PieceDaoTest {
+
+    private BoardDao pieceDao = new BoardDao();
 
     @BeforeEach
     public void before() {
-        BoardDao boardDao = new BoardDao();
-        boardDao.removeAll();
+        pieceDao.removeAll();
     }
 
     @Test
     public void saveTest() {
-        BoardDao boardDao = new BoardDao();
-        assertDoesNotThrow(() -> boardDao.saveState(new WhiteTurn()));
+        Board board = new Board(new BoardBuilder());
+        assertDoesNotThrow(() -> pieceDao.saveAllPieces(board.getBoard()));
     }
 
     @Test
     public void saveAndLoadTest() {
-        BoardDao boardDao = new BoardDao();
-        boardDao.saveState(new WhiteTurn());
-        State state = boardDao.loadState();
-        assertThat(state.getTeam()).isEqualTo(Team.WHITE);
+        Board board = new Board(new BoardBuilder());
+        pieceDao.saveAllPieces(board.getBoard());
+        Map<Position, Piece> pieces = pieceDao.loadAllPieces();
+        assertThat(pieces.values().size()).isEqualTo(64);
     }
 }
