+++ src/test/java/chess/dao/ChessGameDaoTest.java
@@ -1,0 +1,32 @@
+package chess.dao;
+
+import chess.domain.Team;
+import chess.domain.state.turn.State;
+import chess.domain.state.turn.WhiteTurn;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+
+public class BoardDaoTest {
+
+    private final ChessGameDao chessGameDao = new ChessGameDao();
+
+    @BeforeEach
+    public void before() {
+        chessGameDao.removeAll();
+    }
+
+    @Test
+    public void saveTest() {
+        assertDoesNotThrow(() -> chessGameDao.saveState(new WhiteTurn()));
+    }
+
+    @Test
+    public void saveAndLoadTest() {
+        chessGameDao.saveState(new WhiteTurn());
+        State state = chessGameDao.loadState();
+        assertThat(state.getTeam()).isEqualTo(Team.WHITE);
+    }
+}
